Execute emulator tasks in main thread
The threading logic in the emulator was resulting in very long inference
times, due to the long `NvDlaSleepMS` calls when waiting for a new task.
Rather than tuning that sleep loop, this commit just removes all of the
multi-threading logic, since this driver can't overlap multiple
inference tasks or interleave DLA tasks with EMU tasks.